pyproject.toml
[tool.poetry]
name = "overtime-accounting"
version = "0.1.0"
description = ""
authors = ["Koresh06 <korets-24@mail.ru>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.112.0"
sqlalchemy = "^2.0.32"
alembic = "^1.13.2"
bcrypt = "^4.2.0"
passlib = "^1.7.4"
python-multipart = "^0.0.9"
psycopg2-binary = "^2.9.9"
pytest = "^8.3.2"
pytest-asyncio = "^0.23.8"
aiofiles = "^24.1.0"
python-jose = "^3.3.0"
python-dotenv = "^1.0.1"
black = "^24.8.0"
uvicorn = "^0.30.6"
asyncpg = "^0.29.0"
jinja2 = "^3.1.4"
httpx = "^0.27.0"
psycopg = "^3.2.1"
jwt = "^1.3.1"
pydantic = {extras = ["email"], version = "^2.9.1"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
    ".", "app",
]
asyncio_mode = "auto"